+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     * Stage.find
     * Stage.wrap
     * stage:set_index
+    * stage:set_log_icon
     * stage:add_room
     * stage:add_interactable_card
     * stage:add_interactable_card_loop